Implementando ao projeto principal main

+++ ClientCRUD.Api/Controllers/CustomerController.cs
@@ -1,26 +1,38 @@
-﻿using ClientCRUD.Domain.Entities;
-using ClientCRUD.Domain.Services;
+﻿using ClientCRUD.Domain.Services;
 using ClientCRUD.Infra.Services;
 using ClientCRUD.Shared.Parameters;
 using ClientCRUD.Shared.Results;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ClientCRUD.Api.Controllers
 {
     [ApiController]
-    [Route("")]
+    [Route("customer")]
+    [Authorize]
     public class CustomerController : ControllerBase
     {
         private readonly ICustomerServices _customerServices;
-        public CustomerController(CustomerServices customerServices)
+        private readonly IUserServices _userServices;
+        public CustomerController(CustomerServices customerServices, UserServices userServices)
         {
             _customerServices = customerServices;
+            _userServices = userServices;
         }
-        [HttpGet("/customer")]
-        public async Task<IActionResult> GetAll() => Ok(await _customerServices.GetCustomers()); //Done
-        [HttpGet("/customer/less")]
+        [HttpGet("")]
+        public async Task<IActionResult> GetAll() //Done
+        {
+            int[] permissions = [1, 2];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
+            return Ok(await _customerServices.GetCustomers());
+        }
+        [HttpGet("less")]
         public async Task<IActionResult> GetAllWithoutDetails() //Done
         {
+            int[] permissions = [1, 2];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var result = await _customerServices.GetCustomers();
             List<CustomerNoDetails> customers = new List<CustomerNoDetails>();
             for(int i = 0; i < result.Count(); i++)
@@ -29,55 +41,68 @@ namespace ClientCRUD.Api.Controllers
             }
             return Ok(customers);
         }
-        [HttpGet("/customer/{id}")]
+        [HttpGet("{id}")]
         public async Task<IActionResult> GetByCode([FromRoute] string id) //Done
         {
+            int[] permissions = [1, 2];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var customer = await _customerServices.GetCustomerById(id);
             if (customer == null)
                 return NotFound($"Not found the id: {id}");
             return Ok(customer);
         }
-        [HttpGet("/customer/{id}/less")]
+        [HttpGet("{id}/less")]
         public async Task<IActionResult> GetByCodeWithoutDetails([FromRoute] string id) //Done
         {
+            int[] permissions = [1, 2];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var result = await _customerServices.GetCustomerById(id);
             if (result == null)
                 return NotFound($"Not found the id: {id}");
             return Ok(CustomerToNoDetails.Convert(result));
         }
-        [HttpDelete("/customer/{id}")]
+        [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteByCode([FromRoute] string id) //Done
         {
+            int[] permissions = [1];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var result = await _customerServices.DeleteCustomerById(id);
             if (!result) 
                 return NotFound($"Not found the id: {id}");
             return Ok($"{id} deleted with success");
         }
-        [HttpPost("/customer")]
-        public async Task<IActionResult> InsertAllValues([FromBody] CustomerCreate newcustomer) //Done
+        [HttpPost("")]
+        public async Task<IActionResult> InsertAllValues([FromBody] CustomerCreate newcustomer)
         {
+            int[] permissions = [1];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var result = await _customerServices.InsertCustomer(CustomerCreateToCustomer.Convert(newcustomer));
             return Ok(result);
         }
-        [HttpPost("/customer/less")]
-        public async Task<IActionResult> InsertMandatoryValues([FromBody] CustomerCreateMinimum newcustomer) //Done
+        [HttpPost("less")]
+        public async Task<IActionResult> InsertMandatoryValues([FromBody] CustomerCreateMinimum newcustomer)
         {
+            int[] permissions = [1];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var result = await _customerServices.InsertCustomer(CustomerCreateMinimunToCustomer.Convert(newcustomer));
             return Ok(result);
         }
-        [HttpPut("/customer")]
+        [HttpPut("")]
         public async Task<IActionResult> Update([FromBody] CustomerUpdate updtcustomer)
         {
+            int[] permissions = [1];
+            if (!await _userServices.VerifyUserAccess(Guid.Parse(HttpContext.User.Claims.First(c => c.Type == "_id").Value), permissions))
+                return StatusCode(403);
             var customer = await _customerServices.CompareOldNewCustomer(updtcustomer);
             if (customer == null)
                 return NotFound($"Not found the id: {updtcustomer.Id}");
             await _customerServices.UpdateCustomer(customer);
             return Ok(await _customerServices.GetCustomerById(customer.Id.ToString()));
-        }
-        [HttpGet("")]
-        public IActionResult Status()
-        {
-            return StatusCode(418, "API Online");
         }
     }
 }
